refactor: more clear db files structure
